third_party/WebKit/LayoutTests/compositing/iframes/iframe-composited-scrolling-hide-and-show.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>

<style>
iframe {
  height: 100px;
  width: 100px;
}
</style>

<script>
  description('This test makes sure hiding and showing a scrollable iframe ' +
              'correctly updates the set of non-fast scrollable rects.');

  onload = function() {
    if (window.internals) {
      var frame = document.getElementById('scrollable-iframe');
      shouldBe('internals.nonFastScrollableRects(document).length', '1');
      frame.style.display = 'none';
      shouldBe('internals.nonFastScrollableRects(document).length', '0');
      frame.style.display = '';
      shouldBe('internals.nonFastScrollableRects(document).length', '1');
    }
  }
</script>

<iframe id="scrollable-iframe" src="resources/subframe.html"></iframe>
